Move all cats type class instances to cats module, and more tests to tests module
Only remaining use of cats in core: cats.data.Xor
