+++ core/src/main/scala/io/circe/cursor/CObject.scala
@@ -1,6 +1,5 @@
 package io.circe.cursor
 
-import cats.Functor
 import io.circe.{ Context, Cursor, Json, JsonObject }
 
 private[circe] case class CObject(
@@ -33,8 +32,6 @@ private[circe] case class CObject(
   }
 
   def withFocus(f: Json => Json): Cursor = copy(focus = f(focus), u = true)
-  def withFocusM[F[_]](f: Json => F[Json])(implicit F: Functor[F]): F[Cursor] =
-    F.map(f(focus))(j => copy(focus = j, u = true))
 
   override def field(k: String): Option[Cursor] = o(k).map { j =>
     copy(focus = j, key = k)
