+++ core/src/main/scala/io/circe/disjunctionCodecs.scala
@@ -1,6 +1,6 @@
 package io.circe
 
-import cats.data.{ Validated, Xor }
+import cats.data.Xor
 
 /**
  * [[Decoder]] and [[Encoder]] instances for disjunction types with reasonable names for the sides.
@@ -8,8 +8,8 @@ import cats.data.{ Validated, Xor }
 object disjunctionCodecs {
   private[this] val leftKey: String = "Left"
   private[this] val rightKey: String = "Right"
-  private[this] val failureKey: String = "Invalid"
-  private[this] val successKey: String = "Valid"
+  private[circe] val failureKey: String = "Invalid"
+  private[circe] val successKey: String = "Valid"
 
   implicit def decodeXor[A, B](implicit
     da: Decoder[A],
@@ -20,11 +20,6 @@ object disjunctionCodecs {
     da: Decoder[A],
     db: Decoder[B]
   ): Decoder[Either[A, B]] = Decoder.decodeEither(leftKey, rightKey)
-
-  implicit def decodeValidated[E, A](implicit
-    de: Decoder[E],
-    da: Decoder[A]
-  ): Decoder[Validated[E, A]] = Decoder.decodeValidated(failureKey, successKey)
 
   implicit def encodeXor[A, B](implicit
     ea: Encoder[A],
@@ -37,10 +32,4 @@ object disjunctionCodecs {
     eb: Encoder[B]
   ): Encoder[Either[A, B]] =
     Encoder.encodeEither(leftKey, rightKey)
-
-  implicit def encodeValidated[E, A](implicit
-    ee: Encoder[E],
-    ea: Encoder[A]
-  ): Encoder[Validated[E, A]] =
-    Encoder.encodeValidated(failureKey, successKey)
 }
