+++ tests/src/test/scala/io/circe/CursorTests.scala
@@ -1,24 +1,25 @@
 package io.circe
 
+import cats.{Functor, Applicative}
 import io.circe.Cats._
 import cats.std.list._
 import io.circe.test.CursorSuite
 
-class CursorTests extends CursorSuite[Cursor] {
+class CursorTests extends CursorSuite[Cursor, Functor] {
   def fromJson(j: Json): Cursor = Cursor(j)
   def top(c: Cursor): Option[Json] = Some(c.top)
   def focus(c: Cursor): Option[Json] = Some(c.focus)
   def fromResult(result: Option[Cursor]): Option[Cursor] = result
 }
 
-class HCursorTests extends CursorSuite[HCursor] {
+class HCursorTests extends CursorSuite[HCursor, Functor] {
   def fromJson(j: Json): HCursor = Cursor(j).hcursor
   def top(c: HCursor): Option[Json] = Some(c.top)
   def focus(c: HCursor): Option[Json] = Some(c.focus)
   def fromResult(result: ACursor): Option[HCursor] = result.success
 }
 
-class ACursorTests extends CursorSuite[ACursor] {
+class ACursorTests extends CursorSuite[ACursor, Applicative] {
   def fromJson(j: Json): ACursor = Cursor(j).hcursor.acursor
   def top(c: ACursor): Option[Json] = c.top
   def focus(c: ACursor): Option[Json] = c.focus
