+++ tests/src/test/scala/io/circe/test/CirceSuite.scala
